Show All, Grad Search, and Tables
Added a few guides to help the team catch up with adding a show all button that basically searches for anything, and then adding logic to the student controller and search forms to search by major and graduation date. Additionally, there is a screenshot of an example of what the table formatting looks like before any bootstrap formatting.